refactor: 레코드 메서드 dto로 분리
