+++ src/main/java/com/nhnacademy/order_server/service/OrderService.java
@@ -1,14 +1,12 @@
 package com.nhnacademy.order_server.service;
 
 
+import com.nhnacademy.order_server.dto.OrderCalculationData;
 import com.nhnacademy.order_server.dto.message.PaymentSuccessMessage;
 import com.nhnacademy.order_server.dto.request.OrderCreateRequest;
-import com.nhnacademy.order_server.dto.request.OrderReturnRequest;
 import com.nhnacademy.order_server.dto.response.OrderCreateResponse;
 import com.nhnacademy.order_server.dto.response.OrderResponse;
-import com.nhnacademy.order_server.dto.response.OrderReturnCheckResponse;
 import com.nhnacademy.order_server.dto.response.OrderValidationInfoResponse;
-import com.nhnacademy.order_server.service.impl.OrderServiceImpl;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
@@ -27,8 +25,7 @@ public interface OrderService {
             Long userId,
             int usedPoint,
             String orderKey,
-            double earnRate,
-            OrderServiceImpl.OrderCalculationData orderData,
+            OrderCalculationData orderData,
             OrderCreateRequest.OrderCalculationResult calculationResult
     );
 }
